Converts shimmered details list as well
